[Redesign] Addressing the second round of feedback
