+++ wormhole-connect/src/views/v2/TxHistory/Widget/Item.tsx
@@ -1,5 +1,13 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, {
+  useCallback,
+  useContext,
+  useEffect,
+  useMemo,
+  useState,
+} from 'react';
+import { useDispatch } from 'react-redux';
 import { useTimer } from 'react-timer-hook';
+import { useTheme } from '@mui/material';
 import Box from '@mui/material/Box';
 import Card from '@mui/material/Card';
 import CardActionArea from '@mui/material/CardActionArea';
@@ -10,12 +18,21 @@ import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import { makeStyles } from 'tss-react/mui';
 
-import config from 'config';
+import AlertBannerV2 from 'components/v2/AlertBanner';
+import config, { getWormholeContextV2 } from 'config';
+import { RouteContext } from 'contexts/RouteContext';
 import useTrackTransferInProgress from 'hooks/useTrackTransferInProgress';
-import useResumeTransaction from 'hooks/useResumeTransaction';
 import ArrowRight from 'icons/ArrowRight';
 import TokenIcon from 'icons/TokenIcons';
 import TxCompleteIcon from 'icons/TxComplete';
+import {
+  setRoute as setRedeemRoute,
+  setIsResumeTx,
+  setTimestamp,
+  setTxDetails,
+} from 'store/redeem';
+import { setRoute as setAppRoute } from 'store/router';
+import { setToChain } from 'store/transferInput';
 import { toFixedDecimals } from 'utils/balance';
 import { removeTxFromLocalStorage } from 'utils/inProgressTxCache';
 import { minutesAndSecondsWithPadding } from 'utils/transferValidation';
@@ -23,6 +40,9 @@ import { minutesAndSecondsWithPadding } from 'utils/transferValidation';
 import type { TransactionLocal } from 'config/types';
 
 const useStyles = makeStyles()((theme: any) => ({
+  alertBanner: {
+    marginTop: '12px',
+  },
   arrowIcon: {
     fontSize: '16px',
     margin: '0 4px',
@@ -60,12 +80,13 @@ type Props = {
 };
 
 const WidgetItem = (props: Props) => {
+  const [error, setError] = useState('');
   const [etaExpired, setEtaExpired] = useState(false);
-  const [resumeTxLocal, setResumeTxLocal] = useState<
-    TransactionLocal | undefined
-  >();
 
   const { classes } = useStyles();
+  const dispatch = useDispatch();
+  const routeContext = useContext(RouteContext);
+  const theme = useTheme();
 
   const { data: transaction } = props;
   const { receipt, route, timestamp, txDetails, txHash } = transaction;
@@ -83,8 +104,6 @@ const WidgetItem = (props: Props) => {
     receipt,
     route,
   });
-
-  const { isLoading: isLoadingResume } = useResumeTransaction(resumeTxLocal);
 
   useEffect(() => {
     if (isCompleted && txHash) {
@@ -167,6 +186,35 @@ const WidgetItem = (props: Props) => {
     }
   }, [etaRemaining, isCompleted, isRunning]);
 
+  // Action handler to navigate user to the Redeem view of this transaction
+  const resumeTransaction = useCallback(async () => {
+    // Clear previous errors when user clicks on the widget again
+    setError('');
+
+    try {
+      const wh = await getWormholeContextV2();
+      const sdkRoute = new (config.routes.get(route).rc)(wh);
+
+      // Set the start time of the transaction
+      dispatch(setTimestamp(timestamp));
+      // Set transaction details required to display Redeem view
+      dispatch(setTxDetails(txDetails));
+      // Set to avoid send transfer.success event in Resume Transaction case
+      dispatch(setIsResumeTx(true));
+      // Set transaction route
+      dispatch(setRedeemRoute(route));
+      // Set transaction destination chain
+      dispatch(setToChain(txDetails.toChain));
+      // Set the App route to navigate user to Redeem view
+      dispatch(setAppRoute('redeem'));
+
+      routeContext.setRoute(sdkRoute);
+      routeContext.setReceipt(receipt);
+    } catch (e: unknown) {
+      setError(`Error resuming transaction: ${txDetails.sendTx}`);
+    }
+  }, [receipt, route, timestamp, txDetails]);
+
   if (!transaction) {
     return <></>;
   }
@@ -176,17 +224,9 @@ const WidgetItem = (props: Props) => {
       <Card className={classes.card}>
         <CardActionArea
           disableTouchRipple
-          disabled={isLoadingResume || !txDetails}
+          disabled={!txDetails}
           className={classes.cardActionArea}
-          onClick={async () => {
-            setResumeTxLocal({
-              txDetails,
-              timestamp,
-              route,
-              receipt,
-              txHash,
-            });
-          }}
+          onClick={resumeTransaction}
         >
           <CardContent>
             <Stack
@@ -229,6 +269,13 @@ const WidgetItem = (props: Props) => {
           </CardContent>
         </CardActionArea>
       </Card>
+      <AlertBannerV2
+        className={classes.alertBanner}
+        color={error ? theme.palette.error.light : theme.palette.grey.A400}
+        content={error}
+        error={!!error}
+        show={!!error}
+      />
     </div>
   );
 };
